+++ project-mongo/src/main/java/cn/taike/mongo/database/jdbc/DataSourceConfig.java
@@ -1,14 +1,4 @@
-package cn.taike.mongo.dao;
-
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.boot.context.properties.EnableConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Primary;
-
-import javax.sql.DataSource;
+package cn.taike.mongo.database.jdbc;
 
 /**
  * Created by huayandong on 17/3/13.
