+++ project-mongo/src/main/java/cn/taike/mongo/database/jdbc/JdbcTemplateSample.java
@@ -1,4 +1,4 @@
-package cn.taike.mongo.dao;
+package cn.taike.mongo.database.jdbc;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.jdbc.core.JdbcTemplate;
