+++ project-mongo/src/main/java/cn/taike/mongo/database/mysql/BookSectionJpaRepository.java
@@ -1,9 +1,8 @@
-package cn.taike.mongo.mysql;
+package cn.taike.mongo.database.mysql;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
 
 import javax.transaction.Transactional;
 import java.util.List;
