--- project-mongo/src/main/java/cn/taike/mongo/mysql/BookSection.java
@@ -1,75 +1,0 @@
-package cn.taike.mongo.mysql;
-
-import javax.persistence.*;
-
-/**
- * Created by huayandong on 17/7/27.
- */
-@Entity
-@Table(name = "book")
-public class BookSection {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-
-    @Column(name = "book_name")
-    private String bookName;
-
-    @Column(name = "project_name")
-    private String projectName;
-
-    private String type;
-
-    private String cover;
-
-    private String video;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getBookName() {
-        return bookName;
-    }
-
-    public void setBookName(String bookName) {
-        this.bookName = bookName;
-    }
-
-    public String getProjectName() {
-        return projectName;
-    }
-
-    public void setProjectName(String projectName) {
-        this.projectName = projectName;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public String getCover() {
-        return cover;
-    }
-
-    public void setCover(String cover) {
-        this.cover = cover;
-    }
-
-    public String getVideo() {
-        return video;
-    }
-
-    public void setVideo(String video) {
-        this.video = video;
-    }
-}
