--- project-mongo/src/main/java/cn/taike/mongo/syj/BookSectionLabel.java
@@ -1,50 +1,0 @@
-package cn.taike.mongo.syj;
-
-import javax.persistence.*;
-
-/**
- * Created by huayandong on 17/7/27.
- */
-@Entity
-@Table(name = "book_section_label")
-public class BookSectionLabel {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.AUTO)
-    private Integer id;
-    private String lesson;
-    private String level;
-    private String type;
-
-    public Integer getId() {
-        return id;
-    }
-
-    public void setId(Integer id) {
-        this.id = id;
-    }
-
-    public String getLesson() {
-        return lesson;
-    }
-
-    public void setLesson(String lesson) {
-        this.lesson = lesson;
-    }
-
-    public String getLevel() {
-        return level;
-    }
-
-    public void setLevel(String level) {
-        this.level = level;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-}
