--- project-mongo/src/main/java/cn/taike/mongo/yun/CourseContentLabelJpaRepository.java
@@ -1,9 +1,0 @@
-package cn.taike.mongo.yun;
-
-import org.springframework.data.mongodb.repository.MongoRepository;
-
-/**
- * Created by huayandong on 17/7/26.
- */
-public interface CourseContentLabelJpaRepository extends MongoRepository<CourseContentLabel, String> {
-}
