--- project-mongo/src/test/java/cn/taike/mongo/MongoApplicationTests.java
@@ -1,16 +1,0 @@
-package cn.taike.mongo;
-
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@SpringBootTest
-public class MongoApplicationTests {
-
-	@Test
-	public void contextLoads() {
-	}
-
-}
